+++ src/main/java/com/olgaboiar/mint/Body.java
@@ -1,0 +1,9 @@
+package com.olgaboiar.mint;
+
+public class Body {
+    String body;
+
+    Body(String body) {
+        this.body = body;
+    }
+}
