+++ src/main/java/com/olgaboiar/mint/Header.java
@@ -1,0 +1,21 @@
+package com.olgaboiar.mint;
+
+public class Header {
+    private String statusLine;
+    private String contentType;
+    private String blankLine;
+
+    public Header(String statusLine, String contentType) {
+        this.statusLine = statusLine;
+        this.contentType = contentType;
+        this.blankLine = ("");
+    }
+
+    String getStatusLine() {
+        return statusLine;
+    }
+
+    String getContentType() {
+        return contentType;
+    }
+}
