+++ src/main/java/com/olgaboiar/mint/Response.java
@@ -2,22 +2,26 @@ package com.olgaboiar.mint;
 
 public class Response {
 
-    private String statusLine;
-    private String contentType;
-    private String blankLine;
+//    private String statusLine;
+//    private String contentType;
+//    private String blankLine;
+    private Header header;
+    private Body body;
 
-    public Response(String statusLine, String contentType) {
-        this.statusLine = statusLine;
-        this.contentType = contentType;
-        this.blankLine = ("");
+    public Response(Header header, Body body) {
+//        this.statusLine = statusLine;
+//        this.contentType = contentType;
+//        this.blankLine = ("");
+        this.header = header;
+        this.body = body;
     }
 
-    String getStatusLine() {
-        return statusLine;
+    Header getHeader() {
+        return header;
     }
-
-    String getContentType() {
-        return contentType;
-    }
+//
+//    String getContentType() {
+//        return contentType;
+//    }
 
 }
